Issue #2: Refactored application header (extracted components, created entity AppMenu).
