+++ src/components/App/AppDrawer/AppDrawer.module.scss
@@ -1,0 +1,8 @@
+.drawTitle:global(.MuiTypography-root) {
+  display: flex;
+  align-items: center;
+  margin: 0 16px;
+  justify-content: center;
+  font-size: 18px;
+  line-height: 1;
+}
