+++ src/components/App/AppHeader/AppHeader.module.scss
@@ -2,39 +2,3 @@
   position: relative !important;
   // min-height: 56px;
 }
-
-.navButtons {
-  gap: 8px;
-  :global(.MuiButton-startIcon) {
-    opacity: 0.5;
-  }
-  :global(.MuiButton-root) {
-    transition: opacity $animationTimeMs;
-    :global(.theme-mode-light) &:hover {
-      opacity: 0.7;
-    }
-  }
-}
-
-.appTitle {
-  &:global(.MuiTypography-root) {
-    font-size: 18px;
-    font-weight: normal;
-  }
-}
-
-.ListItemIcon {
-  &:global(.MuiListItemIcon-root) {
-    color: $primaryColor;
-    min-width: 40px;
-  }
-}
-
-.drawTitle:global(.MuiTypography-root) {
-  display: flex;
-  align-items: center;
-  margin: 0 16px;
-  justify-content: center;
-  font-size: 18px;
-  line-height: 1;
-}
