+++ src/components/App/AppTitleBlock/AppTitleBlock.module.scss
@@ -1,0 +1,6 @@
+.appTitle {
+  &:global(.MuiTypography-root) {
+    font-size: 18px;
+    font-weight: normal;
+  }
+}
