+++ src/components/App/AppTitleBlock/index.ts
@@ -1,0 +1,1 @@
+export * from './AppTitleBlock';
