+++ src/entities/AppMenu/components/AppMenu/AppMenu.module.scss
@@ -1,0 +1,12 @@
+.navButtons {
+  gap: 8px;
+  :global(.MuiButton-startIcon) {
+    opacity: 0.5;
+  }
+  :global(.MuiButton-root) {
+    transition: opacity $animationTimeMs;
+    :global(.theme-mode-light) &:hover {
+      opacity: 0.7;
+    }
+  }
+}
