+++ src/entities/AppMenu/components/AppMenu/index.ts
@@ -1,0 +1,1 @@
+export * from './AppMenu';
