+++ src/entities/AppMenu/components/AppSideMenu/AppSideMenu.module.scss
@@ -1,0 +1,6 @@
+.ListItemIcon {
+  &:global(.MuiListItemIcon-root) {
+    color: $primaryColor;
+    min-width: 40px;
+  }
+}
