+++ src/entities/AppMenu/components/AppSideMenu/index.ts
@@ -1,0 +1,1 @@
+export * from './AppSideMenu';
