+++ src/entities/AppMenu/components/index.ts
@@ -1,0 +1,2 @@
+export * from './AppMenu';
+export * from './AppSideMenu';
