+++ src/entities/AppMenu/constants/AppMenuConstants.ts
@@ -1,0 +1,2 @@
+/** The width of mobile menu drawer */
+export const drawerWidth = 280;
