+++ src/entities/AppMenu/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useAppMenuItems';
+export * from './useHandleAppMenuClick';
