+++ src/entities/AppMenu/index.ts
@@ -1,0 +1,4 @@
+export * from './components';
+export * from './constants';
+export * from './hooks';
+export * from './types';
