+++ src/entities/AppMenu/types/index.ts
@@ -1,0 +1,1 @@
+export * from './AppMenuTypes';
